+++ CHANGES.rst
@@ -11,6 +11,7 @@ Changelog
 - added `-p/--pretty-print` flag to `to-alpaca` writer
 - added `pairs-to-llama2` and `llama2-to-pairs` filter
   (since llama2 has instruction as part of the string, it is treated as pretrain data)
+- added `to-llama2-format` filter for pretrain records (no [INST]...[/INST] block)
 
 
 0.0.2 (2023-10-31)